+++ packages/forma-36-website/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.16](https://github.com/contentful/forma-36/compare/forma-36-website@0.5.15...forma-36-website@0.5.16) (2020-05-29)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.5.15](https://github.com/contentful/forma-36/compare/forma-36-website@0.5.14...forma-36-website@0.5.15) (2020-05-29)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.5.14](https://github.com/contentful/forma-36/compare/forma-36-website@0.5.13...forma-36-website@0.5.14) (2020-05-26)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.5.13](https://github.com/contentful/forma-36/compare/forma-36-website@0.5.12...forma-36-website@0.5.13) (2020-05-20)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
+## [0.5.12](https://github.com/contentful/forma-36/compare/forma-36-website@0.5.11...forma-36-website@0.5.12) (2020-05-19)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
 ## [0.5.11](https://github.com/contentful/forma-36/compare/forma-36-website@0.5.10...forma-36-website@0.5.11) (2020-05-19)
 
 **Note:** Version bump only for package forma-36-website